requirements.txt
Django==4.2
environs[django]==9.5.0
django-debug-toolbar==4.0.0
gunicorn==20.0.2
psycopg2-binary==2.9.5
django-widget-tweaks==1.4.12
Pillow==9.5.0
django-apscheduler==0.6.2
python-dateutil==2.8.2
yookassa==2.3.5
qrcode==7.4.2
